[medUtilities] seriesDicomId (don't merge yet)
I got the duplication/overwriting bug we had a while back while importing data. The function ```generateSeriesAndSOPInstanceId``` fixed this before, but I got the bug on the output of registration. I corrected this by generating a new series DICOM ID in addition to the series ID that is already generated.
This PR shouldn't be merged yet because the function name should be changed to reflect this, but I'll wait for later to do the change (since it will affect all our plugins). I wanted to post an issue but I can't on this repo.